+++ src/modules/MediaGrid/MediaGrid.tsx
@@ -1,8 +1,84 @@
-import { $, component$, QRL, useSignal } from "@builder.io/qwik";
-import { MediaBase } from "~/services/types3";
-import { MediaCard } from "../MediaCard/MediaCard";
+import { $, component$, QRL, useComputed$, useSignal } from "@builder.io/qwik";
+import { Stars } from "~/components/Stars/Stars";
+import { getPoster, getPosterSet } from "~/services/images";
+import { MediaBase } from "~/services/types";
+import { paths } from "~/utils/paths";
+import { getHeading, getMediaType } from "./MediaCard.utils";
 
-type Props = {
+type MediaCardProps = {
+  media: MediaBase;
+};
+
+export const MediaCard = component$((props: MediaCardProps) => {
+  const mediaType = useComputed$(() => {
+    return getMediaType(props.media);
+  });
+
+  const heading = useComputed$(() => {
+    return getHeading(props.media);
+  });
+
+  return (
+    <a href={paths.media(mediaType.value, props.media.id)} class="w-48">
+      <div class="transition-scale scale-95 duration-300 ease-in-out hover:scale-100">
+        <picture>
+          <img
+            alt={heading.value}
+            class="max-w-full border-4 border-base-300 object-cover "
+            height={270}
+            src={getPoster(props.media, "92")}
+            srcSet={getPosterSet(props.media, "185")}
+            width={185}
+          />
+        </picture>
+      </div>
+      <span>{heading.value}</span>
+      <Stars rating={props.media.vote_average} />
+    </a>
+  );
+});
+
+type MediaCarouselProps = {
+  collection: MediaBase[];
+  title: string;
+  viewAllHref: string;
+};
+
+export const MediaCarousel = component$((props: MediaCarouselProps) => {
+  return (
+    <section>
+      <div class="flex flex-row items-center px-12 py-2">
+        <h2 class="text-2xl text-white">{props.title}</h2>
+        <div class="flex-auto" />
+        <a
+          class="transition-text opacity-80 duration-100 ease-in-out hover:text-qwik-light-blue hover:opacity-100"
+          href={props.viewAllHref}
+        >
+          Explore All
+        </a>
+      </div>
+      <div class="relative">
+        <div class="overflow-y-auto px-8 py-4">
+          <div class="carousel flex w-max flex-row gap-2">
+            {props.collection?.map((media) => (
+              <div class="carousel-item" key={media.id}>
+                <MediaCard media={media} />
+              </div>
+            ))}
+            <a
+              class="transition-text flex w-44 items-center justify-center duration-100 ease-in-out hover:text-qwik-light-blue"
+              href={props.viewAllHref}
+            >
+              Explore All
+            </a>
+          </div>
+        </div>
+      </div>
+    </section>
+  );
+});
+
+type MediaGridProps = {
   collection: MediaBase[];
   currentPage: number;
   onMore$?: QRL<() => void>;
@@ -10,7 +86,7 @@ type Props = {
   parentContainer?: Element | null;
 };
 
-export const MediaGrid = component$((props: Props) => {
+export const MediaGrid = component$((props: MediaGridProps) => {
   const throttleTimer = useSignal(false);
   const scrollEnabled = useSignal(props.currentPage < props.pageCount);
