+++ src/modules/PersonHero/PersonHero.tsx
@@ -1,6 +1,6 @@
 import { component$ } from "@builder.io/qwik";
 import { getProfile, getProfileSet } from "~/services/images";
-import type { PersonMediaDetails } from "~/services/types";
+import { PersonDetails } from "~/services/types";
 import { formatDate } from "~/utils/format";
 import { ExternalLinks } from "../ExternalLinks/ExternalLinks";
 
@@ -12,7 +12,7 @@ export const calculateAge = (birthday: string, deathday?: string) => {
 };
 
 type Props = {
-  person: PersonMediaDetails;
+  person: PersonDetails;
 };
 
 export const PersonHero = component$((props: Props) => {
@@ -96,7 +96,7 @@ export const PersonHero = component$((props: Props) => {
 
           <div>
             <ExternalLinks
-              media="person"
+              isPerson
               links={{
                 ...props.person.external_ids,
                 homepage: props.person.homepage,
