+++ src/routes/genre/[genreId]/movie/index.tsx
@@ -2,7 +2,7 @@ import { component$, useSignal } from "@builder.io/qwik";
 import { DocumentHead, routeLoader$, server$, z } from "@builder.io/qwik-city";
 import { MediaGrid } from "~/modules/MediaGrid/MediaGrid";
 import { getMediaByGenre, getTMDBContext } from "~/services/tmdb";
-import { MediaBase } from "~/services/types3";
+import { MediaBase } from "~/services/types";
 import { paths } from "~/utils/paths";
 
 export const useGenreMovies = routeLoader$(async (event) => {
@@ -58,7 +58,6 @@ export default component$(() => {
       <h1 class="px-8 pt-4 text-4xl">{`Movie Genre: ${
         movies?.value.genre?.name || "Not defined"
       }`}</h1>
-      <pre>{JSON.stringify(movies.value, null, 2)}</pre>
       <MediaGrid
         collection={collection.value}
         currentPage={currentPage.value}
