+++ src/routes/movie/(all)/index.tsx
@@ -1,7 +1,7 @@
 import { component$ } from "@builder.io/qwik";
 import { routeLoader$, type DocumentHead } from "@builder.io/qwik-city";
 import { Footer } from "~/modules/Footer/Footer";
-import { MediaCarousel } from "~/modules/MediaCarousel/MediaCarousel";
+import { MediaCarousel } from "~/modules/MediaGrid/MediaGrid";
 import { MovieHero } from "~/modules/MovieHero/MovieHero";
 import {
   getMovie,
