+++ src/services/types.ts
@@ -1,64 +1,4 @@
-type TvMedia = {
-  backdrop_path?: string | null;
-  first_air_date?: string[];
-  genre_ids?: number[];
-  id: number;
-  media_type?: "tv";
-  name?: string;
-  origin_country?: string[];
-  original_language?: string;
-  original_name?: string;
-  overview?: string[];
-  popularity?: number;
-  poster_path?: string | null;
-  vote_average?: number;
-  vote_count?: number;
-};
-
-type MovieMedia = {
-  adult?: boolean;
-  backdrop_path?: string | null;
-  genre_ids?: number[];
-  id: number;
-  media_type?: "movie";
-  original_language?: string;
-  original_title?: string;
-  overview?: string;
-  popularity?: number;
-  poster_path?: string | null;
-  release_date?: string;
-  title?: string;
-  video?: boolean;
-  vote_average?: number;
-  vote_count?: number;
-};
-
-type ProductionMedia = TvMedia | MovieMedia;
-
-export type PersonMedia = {
-  profile_path?: string;
-  adult?: boolean;
-  id: number;
-  known_for?: ProductionMedia[];
-  name?: string;
-  popularity?: number;
-  media_type?: "person";
-  job?: string;
-};
-
-export type PersonMediaDetails = PersonMedia & {
-  external_ids?: Record<string, string>;
-  homepage?: string;
-  biography?: string;
-  known_for_department?: string;
-  birthday?: string;
-  deathday?: string;
-  place_of_birth?: string;
-  combined_credits?: {
-    cast: ProductionMedia[];
-    crew: ProductionMedia[];
-  };
-};
+/* eslint-disable @typescript-eslint/no-explicit-any */
 
 export type MediaType = "movie" | "tv" | "person";
 
@@ -68,3 +8,309 @@ export type Collection<T> = {
   total_pages?: number;
   total_results?: number;
 };
+
+export type Genre = {
+  id: number;
+  name: string;
+};
+
+export type ProductionCompany = {
+  id: number;
+  logo_path: string;
+  name: string;
+  origin_country: string;
+};
+
+export type ProductionCountry = {
+  iso_3166_1: string;
+  name: string;
+};
+
+export type SpokenLanguage = {
+  english_name: string;
+  iso_639_1: string;
+  name: string;
+};
+
+export type VideoResult = {
+  iso_639_1: string;
+  iso_3166_1: string;
+  name: string;
+  key: string;
+  site: string;
+  size: number;
+  type: string;
+  official: boolean;
+  published_at: string;
+  id: string;
+};
+
+export type Videos = {
+  results: VideoResult[];
+};
+
+export type Cast = {
+  adult: boolean;
+  gender: number;
+  id: number;
+  known_for_department: string;
+  name: string;
+  original_name: string;
+  popularity: number;
+  profile_path?: string;
+  cast_id: number;
+  character: string;
+  credit_id: string;
+  order: number;
+};
+
+export type Crew = {
+  adult: boolean;
+  gender: number;
+  id: number;
+  known_for_department: string;
+  name: string;
+  original_name: string;
+  popularity: number;
+  profile_path?: string;
+  credit_id: string;
+  department: string;
+  job: string;
+};
+
+export type Credits = {
+  cast: Cast[];
+  crew: Crew[];
+};
+
+export type Image = {
+  aspect_ratio: number;
+  height: number;
+  iso_639_1: string;
+  file_path: string;
+  vote_average: number;
+  vote_count: number;
+  width: number;
+};
+
+export type PersonImages = {
+  profiles: Image[];
+};
+
+export type Images = {
+  backdrops: Image[];
+  logos: Image[];
+  posters: Image[];
+};
+
+export type ExternalIds = {
+  imdb_id: string;
+  wikidata_id: string;
+  facebook_id: any;
+  instagram_id: any;
+  twitter_id: any;
+};
+
+export type ReleaseDate = {
+  certification: string;
+  descriptors: any[];
+  iso_639_1: string;
+  note: string;
+  release_date: string;
+  type: number;
+};
+
+export type ReleaseDateResults = {
+  iso_3166_1: string;
+  release_dates: ReleaseDate[];
+};
+
+export type ReleaseDates = {
+  results: ReleaseDateResults[];
+};
+
+export type CreatedBy = {
+  id: number;
+  credit_id: string;
+  name: string;
+  gender: number;
+  profile_path: string;
+};
+
+export type LastEpisodeToAir = {
+  id: number;
+  name: string;
+  overview: string;
+  vote_average: number;
+  vote_count: number;
+  air_date: string;
+  episode_number: number;
+  episode_type: string;
+  production_code: string;
+  runtime: number;
+  season_number: number;
+  show_id: number;
+  still_path: string;
+};
+
+export type NextEpisodeToAir = {
+  id: number;
+  name: string;
+  overview: string;
+  vote_average: number;
+  vote_count: number;
+  air_date: string;
+  episode_number: number;
+  episode_type: string;
+  production_code: string;
+  runtime: any;
+  season_number: number;
+  show_id: number;
+  still_path: string;
+};
+
+export type Network = {
+  id: number;
+  logo_path: string;
+  name: string;
+  origin_country: string;
+};
+
+export type Season = {
+  air_date: string;
+  episode_count: number;
+  id: number;
+  name: string;
+  overview: string;
+  poster_path: string;
+  season_number: number;
+  vote_average: number;
+};
+
+export type ContentRatingsResult = {
+  descriptors: any[];
+  iso_3166_1: string;
+  rating: string;
+};
+
+export type ContentRatings = {
+  results: ContentRatingsResult[];
+};
+
+export type MovieBase = {
+  adult: boolean;
+  backdrop_path: string;
+  id: number;
+  title: string;
+  original_language: string;
+  original_title: string;
+  overview: string;
+  poster_path: string;
+  media_type: "movie";
+  genre_ids?: number[];
+  popularity: number;
+  release_date: string;
+  video: boolean;
+  vote_average: number;
+  vote_count: number;
+};
+
+export type TvBase = {
+  adult?: boolean;
+  backdrop_path: string;
+  id: number;
+  name: string;
+  original_language: string;
+  original_name: string;
+  overview: string;
+  poster_path: string;
+  media_type: "tv";
+  genre_ids?: number[];
+  popularity: number;
+  first_air_date: string;
+  vote_average: number;
+  vote_count: number;
+  origin_country: string[];
+};
+
+export type MediaBase = MovieBase | TvBase;
+
+export type MovieDetails = MovieBase & {
+  belongs_to_collection: any;
+  budget: number;
+  genres: Genre[];
+  homepage: string;
+  imdb_id: string;
+  production_companies: ProductionCompany[];
+  production_countries: ProductionCountry[];
+  revenue: number;
+  runtime: number;
+  spoken_languages: SpokenLanguage[];
+  status: string;
+  tagline: string;
+};
+
+export type TvDetails = TvBase & {
+  created_by: CreatedBy[];
+  episode_run_time: number[];
+  genres: Genre[];
+  homepage: string;
+  in_production: boolean;
+  languages: string[];
+  last_air_date: string;
+  last_episode_to_air: LastEpisodeToAir;
+  next_episode_to_air: NextEpisodeToAir;
+  networks: Network[];
+  number_of_episodes: number;
+  number_of_seasons: number;
+  production_companies: ProductionCompany[];
+  production_countries: ProductionCountry[];
+  seasons: Season[];
+  spoken_languages: SpokenLanguage[];
+  status: string;
+  tagline: string;
+  type: string;
+};
+
+export type MediaDetails = MovieDetails | TvDetails;
+
+export type MovieExtraDetails = MovieDetails & {
+  videos: Videos;
+  credits: Credits;
+  images: Images;
+  external_ids: ExternalIds;
+  release_dates: ReleaseDates;
+};
+
+export type TvExtraDetails = TvDetails & {
+  videos: Videos;
+  credits: Credits;
+  images: Images;
+  external_ids: ExternalIds;
+  content_ratings: ContentRatings;
+};
+
+export type MediaExtraDetails = MovieExtraDetails | TvExtraDetails;
+
+export type PersonDetails = {
+  adult: boolean;
+  also_known_as: string[];
+  biography: string;
+  birthday: string;
+  deathday: any;
+  gender: number;
+  homepage: any;
+  id: number;
+  imdb_id: string;
+  known_for_department: string;
+  name: string;
+  place_of_birth: string;
+  popularity: number;
+  profile_path: string;
+  images: PersonImages;
+  combined_credits: Credits;
+  external_ids: ExternalIds;
+  media_type: "person";
+};
+
+export type Person = Cast | Crew | PersonDetails;
